Change reports controller
